--- a/src/functional/attention.jl
+++ b/src/functional/attention.jl
@@ -20,18 +20,11 @@ end
     q::CollapsedDimsArray, k::CollapsedDimsArray, v::CollapsedDimsArray,
     args...
 )
-    hq = (move_head_dim_out ∘ split_head)(head, q)
-    hk = (move_head_dim_out ∘ split_head)(head, k)
-    hv = (move_head_dim_out ∘ split_head)(head, v)
+    hq = _split_and_move_head(head, q)
+    hk = _split_and_move_head(head, k)
+    hv = _split_and_move_head(head, v)
     t = generic_qkv_attention(mixingf, scoref, hq, hk, hv, args...)
-    if t isa NamedTuple
-        a = t.hidden_state
-        y = (merge_head ∘ move_head_dim_in)(a)
-        return merge(t, (hidden_state = y,))
-    else
-        a = t
-        return (merge_head ∘ move_head_dim_in)(a)
-    end
+    return _move_and_merge_head(t)
 end
 
 naive_attention_score() = normalized_score(NNlib.softmax) $ scaled_dot_product_score