Test script:
```julia
using CUDA, Zygote, NNlib, ChainRulesCore, NeuralAttentionlib, Flux
using NeuralAttentionlib.Matmul, NeuralAttentionlib.Masks

include("test/old_impl/old_impl.jl")
using .Old_Impl

function atten(mha, x, mask = nothing, p = nothing)
    q = mha.iqproj(x)
    k = mha.ikproj(x)
    v = mha.ivproj(x)
    a = NeuralAttentionlib.multihead_qkv_attention(4, q, k, v, mask, p)
    return mha.oproj(a)
end

head = 4
input_dims = 7
head_dims = 5
output_dims = 8
mha = MultiheadAttention(head, input_dims, head_dims, output_dims; future = false, pdrop = 0);

q_len = [3, 5]
k_len = [4, 4]
lmask = RevBiLengthMask(q_len, k_len)
mask = CausalMask() & lmask

old_q_len = reverse!(reverse!(Old_Impl.getmask(map(l->ones(l), q_len)); dims=1); dims=2)
old_k_len = reverse!(reverse!(Old_Impl.getmask(map(l->ones(l), k_len)); dims=1); dims=2)
old_mask = zeros(6, 6, 2)
old_mask[3:6, 2:6, :] .= Old_Impl.getmask(old_k_len, old_q_len)

mha = gpu(mha);
mask = gpu(mask);
old_mask = gpu(old_mask);
x = cu(randn(input_dims,6,2));
```

and the results:

```jl
# before PR
julia> @btime CUDA.@sync Flux.gradient($mha, $x, nothing) do mha, x, m
           sum(sin.(atten(mha, x, m)))
       end;
  2.477 ms (6906 allocations: 381.14 KiB)

julia> @btime CUDA.@sync Flux.gradient($mha, $x, $mask) do mha, x, m
           sum(sin.(atten(mha, x, BatchedMask(m))))
       end;
  2.620 ms (7387 allocations: 439.06 KiB)

julia> @btime CUDA.@sync Flux.gradient($mha, $x, $mask) do mha, x, m
           sum(sin.(atten(mha, x, BatchedMask(m), 0.1)))
       end;
  2.704 ms (7646 allocations: 450.42 KiB)

# after PR
julia> @btime CUDA.@sync Flux.gradient($mha, $x, nothing) do mha, x, m
           sum(sin.(atten(mha, x, BatchedMask(m))))
       end;
  495.150 μs (1314 allocations: 134.94 KiB)

julia> @btime CUDA.@sync Flux.gradient($mha, $x, $mask) do mha, x, m
           sum(sin.(atten(mha, x, BatchedMask(m))))
       end;
  528.398 μs (1430 allocations: 150.34 KiB)

julia> @btime CUDA.@sync Flux.gradient($mha, $x, $mask) do mha, x, m
           sum(sin.(atten(mha, x, BatchedMask(m), 0.1)))
       end;
  563.092 μs (1549 allocations: 159.55 KiB)

# compare to old impl
julia> @btime CUDA.@sync Flux.gradient($mha, $x, nothing) do mha, x, m
           sum(sin.(mha(x,x,x, mask=m)))
       end;
  693.259 μs (1910 allocations: 102.03 KiB)

julia> @btime CUDA.@sync Flux.gradient($mha, $x, $old_mask) do mha, x, m
           sum(sin.(mha(x,x,x, mask=m)))
       end;
  705.395 μs (1961 allocations: 105.02 KiB)


```